Fix product and follow racket conventions
The absorption loop was buggy, as it allowed for (* -1 +inf.0) = +inf.0 and related bad cases.
It used to override Racket defaults to ensure that (* 0 +inf.0) = +nan.0, but since this was done
in the absorption loop, this behaviour is gone.
It seems more reasonable and maybe more predictable to use the same conventions as Racket
(and racket-cas I guess).